Enable authorization for this tool
You should be able to restrict access to this tool. Former `Authorize` middleware bypassed any authorization check. 

See https://nova.laravel.com/docs/4.0/customization/tools.html#authorization on how to configure authorization for tools.

This change also affects the card which is part of this package as it uses routes using the `Authorize` middleware.
So if you want to use the card, make sure that it uses the same `canSee` as the tool.